requirements.txt
fastapi
uvicorn[standard]
pandas
pyyaml
python-dateutil
pydantic

# LLM
groq

# RAG
sentence-transformers
faiss-cpu
pypdf

# Data generation
numpy
